+++ src/base/window_manager.c
@@ -277,8 +277,6 @@ ret_t window_manager_dispatch_input_event(widget_t* widget, event_t* e) {
   return_value_if_fail(wm->vt->dispatch_input_event != NULL, RET_BAD_PARAMS);
 
   if (wm->ignore_input_events) {
-    log_debug("waiting cursort, ignore input events");
-
     return RET_STOP;
   }